Separando, add ignore e trocando nome MAIN
