+++ bibliotecas/menu.c
@@ -1,0 +1,32 @@
+void menuOrdenacao();
+void menuMetodo();
+
+// Mini interface para Ordenação do Vetor
+void menuOrdenacao()
+{
+  printf("\n");
+  printf("\t\t\t ----------------- \n");
+  printf("\t\t\t|    Ordenacao    |\n");
+  printf("\t\t\t|-----------------|\n");
+  printf("\t\t\t| 1-Aleatoria     |\n");
+  printf("\t\t\t| 2-Crescente     |\n");
+  printf("\t\t\t| 3-Decrescente   |\n");
+  printf("\t\t\t -----------------\n");
+  printf("\n\t\t\t Digite 4 para sair\n");
+}
+
+// Mini interface para Métodos
+void menuMetodo()
+{
+  printf("\t\t\t ----------------- \n");
+  printf("\t\t\t|  Metodos Sort   |\n");
+  printf("\t\t\t|-----------------|\n");
+  printf("\t\t\t| 1-Bubble        |\n");
+  printf("\t\t\t| 2-Selection     |\n");
+  printf("\t\t\t| 3-Insert        |\n");
+  printf("\t\t\t| 4-Quick         |\n");
+  printf("\t\t\t| 5-Merge         |\n");
+  printf("\t\t\t| 6-Heap          |\n");
+  printf("\t\t\t -----------------\n");
+  printf("\n\t\t\t Digite 7 para sair\n");
+}
